Add find_char_forward/backward widget
This commit adds support for finding a char on the line currently being
edited.

find_char_forward is invoked with Ctrl+T.

find_char_backward is invoked with Ctrl+B.

Return finds the next match forward and Ctrl+S does the same backward,
but only when the widget is active. If no further matches are in the
buffer, the cursor will be put at the beginning and end respectively.